Safari-609.1.11.1/LayoutTests/webgl/2.0.0/conformance2/misc/expando-loss-2-expected.txt

CONSOLE MESSAGE: line 50: TypeError: null is not an object (evaluating 'instance.expando1 = expandoValue')
This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.

Test: ../../resources/webgl_test_files/conformance2/misc/expando-loss-2.html
[ 1: FAIL ] successfullyParsed should be true (of type boolean). Was undefined (of type undefined).
[ FAIL ] 1 failures reported
